Set ENABLE_* flags in setup.py
So turns out CMake doesn't actually support arbitrary expressions in set()
commands.  Doing something like

  set(A (B AND C))

will set the value of A to the string "(B AND C)".  The AND operator is only
recognized within the context of an if() command.  Move this logic to Python to
make it make sense.